Refactors Wizard functions into WizardService
