+++ src/main/ngapp/src/app/dataservices/connection-table-selector/connection-table-selector.component.spec.ts
@@ -8,11 +8,10 @@ import { AppSettingsService } from "@core/app-settings.service";
 import { LoggerService } from "@core/logger.service";
 import { JdbcTableSelectorComponent } from "@dataservices/jdbc-table-selector/jdbc-table-selector.component";
 import { SelectedTableComponent } from "@dataservices/selected-table/selected-table.component";
-import { DataserviceService } from "@dataservices/shared/dataservice.service";
-import { MockDataserviceService } from "@dataservices/shared/mock-dataservice.service";
 import { MockVdbService } from "@dataservices/shared/mock-vdb.service";
 import { NotifierService } from "@dataservices/shared/notifier.service";
 import { VdbService } from "@dataservices/shared/vdb.service";
+import { WizardService } from "@dataservices/shared/wizard.service";
 import { NgxDatatableModule } from "@swimlane/ngx-datatable";
 import { ConnectionTableSelectorComponent } from "./connection-table-selector.component";
 
@@ -25,9 +24,8 @@ describe("ConnectionTableSelectorComponent", () => {
       imports: [ FormsModule, HttpModule, NgxDatatableModule ],
       declarations: [ ConnectionTableSelectorComponent, JdbcTableSelectorComponent, SelectedTableComponent ],
       providers: [
-        AppSettingsService, LoggerService, NotifierService,
+        AppSettingsService, LoggerService, NotifierService, WizardService,
         { provide: ConnectionService, useClass: MockConnectionService },
-        { provide: DataserviceService, useClass: MockDataserviceService },
         { provide: VdbService, useClass: MockVdbService }
       ]
     })
