+++ src/main/ngapp/src/app/dataservices/jdbc-table-selector/jdbc-table-selector.component.spec.ts
@@ -7,11 +7,10 @@ import { MockConnectionService } from "@connections/shared/mock-connection.servi
 import { AppSettingsService } from "@core/app-settings.service";
 import { LoggerService } from "@core/logger.service";
 import { SelectedTableComponent } from "@dataservices/selected-table/selected-table.component";
-import { DataserviceService } from "@dataservices/shared/dataservice.service";
-import { MockDataserviceService } from "@dataservices/shared/mock-dataservice.service";
 import { MockVdbService } from "@dataservices/shared/mock-vdb.service";
 import { NotifierService } from "@dataservices/shared/notifier.service";
 import { VdbService } from "@dataservices/shared/vdb.service";
+import { WizardService } from "@dataservices/shared/wizard.service";
 import { JdbcTableSelectorComponent } from "./jdbc-table-selector.component";
 
 describe("JdbcTableSelectorComponent", () => {
@@ -23,9 +22,8 @@ describe("JdbcTableSelectorComponent", () => {
       imports: [ FormsModule, HttpModule ],
       declarations: [ JdbcTableSelectorComponent, SelectedTableComponent ],
       providers: [
-        AppSettingsService, LoggerService, NotifierService,
+        AppSettingsService, LoggerService, NotifierService, WizardService,
         { provide: ConnectionService, useClass: MockConnectionService },
-        { provide: DataserviceService, useClass: MockDataserviceService },
         { provide: VdbService, useClass: MockVdbService },
       ]
     })
