+++ src/main/ngapp/src/app/dataservices/shared/wizard.service.spec.ts
@@ -1,0 +1,15 @@
+import { inject, TestBed } from "@angular/core/testing";
+
+import { WizardService } from "./wizard.service";
+
+describe("WizardService", () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [WizardService]
+    });
+  });
+
+  it("should be created", inject([WizardService], (service: WizardService) => {
+    expect(service).toBeTruthy();
+  }));
+});
